# Fixes bufio buffer full error when reading from an array larger than 4096 bytes