I get this issue when retrieving large blobs from [redix](https://github.com/alash3al/redix) which is API compliant with redis.

I created a test that reproduces that error. It happens when reading from a byte array which length is larger than 4096.

See also #641

## Expected Behavior

The blob should be read in full.

## Current Behavior

An error is returned.

## Possible Solution

See, patch in this PR.

## Steps to Reproduce

See the added test case in this PR which (when ran without the fix added) would result in the following:

```bash
❯ go test internal/proto/reader.go internal/proto/reader_test.go 
--- FAIL: TestReader_ReadLine (0.00s)
    reader_test.go:51: Should be able to read the full buffer: bufio: buffer full
    reader_test.go:55: Values must be equal: ""
FAIL
FAIL	command-line-arguments	0.272s
FAIL
```