@eelcocramer 2020-10-09
Changed the implementation.

Original benchmark was:

```
goos: darwin
goarch: amd64
BenchmarkReader_ParseReply_Status-8   	17320110	       66.5 ns/op
BenchmarkReader_ParseReply_Int-8      	33233448	       33.9 ns/op
BenchmarkReader_ParseReply_Error-8    	16776595	       77.0 ns/op
BenchmarkReader_ParseReply_String-8   	13546156	       85.4 ns/op
BenchmarkReader_ParseReply_Slice-8    	3824382	      317 ns/op
PASS
ok  	command-line-arguments	8.036s
```

After the change the benchmark is a bit worse:

````
oos: darwin
goarch: amd64
BenchmarkReader_ParseReply_Status-8   	15872990	       79.6 ns/op
BenchmarkReader_ParseReply_Int-8      	31321155	       37.4 ns/op
BenchmarkReader_ParseReply_Error-8    	16365578	       84.5 ns/op
BenchmarkReader_ParseReply_String-8   	13537149	       92.0 ns/op
BenchmarkReader_ParseReply_Slice-8    	3660882	      331 ns/op
PASS
ok  	command-line-arguments	8.543s
```